权限管理系统：删除菜单/权限

+++ c-core/src/main/java/com/chang/ccloud/service/SysTreeService.java
@@ -19,6 +19,6 @@ public interface SysTreeService {
 
     List<SysMenuDTO> menuTree();
 
-    List<TreeViewDTO> menuToTreeView(List<SysMenuDTO> list, Integer roleId);
+    List<TreeViewDTO> menuToTreeView(List<SysMenuDTO> list, long roleId);
 
 }
